+++ src/Numbers.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import Button from './Button';
+
+function Numbers({ onClick }) {
+  return (
+    <p>
+      {[1, 2, 3, 4, 5].map((i) => (
+        <Button
+          key={i}
+          onClick={() => onClick(i)}
+        >
+          {i}
+        </Button>
+      ))}
+    </p>
+  );
+}
+
+export default Numbers;
