+++ src/Page.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import Counter from './Counter';
+import Numbers from './Numbers';
+
+function Page({ count, onClick }) {
+  return (
+    <div>
+      <p>Counter!</p>
+      <Counter
+        count={count}
+        onClick={onClick}
+      />
+      <Numbers
+        onClick={onClick}
+      />
+    </div>
+  );
+}
+
+export default Page;
